Change from 3D to 2D colliders
Reoriented everything to be about the X, Y axis instead of X, Z.
Encapsulated player, enemy, and shot objects inside of another game
object, as the rotations of the 3D models were not playing nicely with
the 2D planes.
Changed nearly every script to deal well with this.
All old prefabs still exist, with a few modifications. Will remove these
when we're sure everything works, but for now use the prefabs in the
EncapsulatedPrefabs folder.

Things not working: shot collisions, player colliding with enemies,
bullets despawning.