dagger bullet and fireball bullet
Added a dagger bullet and fireball bullet. I also changed their
encapsulation parent objects so that they get destroyed within a certain
time. I created a simple Rotate script that is attached to the dagger
bullet so that while it is moving it rotates in a circle (the child quad
is rotating, not the whole object!).